dracor-mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dracor-mcp-server"
version = "1.0.0"
description = "MCP server for the Drama Corpora Project (DraCor) API - HTTP streaming version for Railway deployment"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "DraCor MCP Team" }
]
keywords = ["mcp", "dracor", "drama", "corpora", "digital-humanities", "literary-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "mcp[cli]>=1.25.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.27.0",
    "starlette>=0.36.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://github.com/dracor-org/dracor-mcp"
Documentation = "https://dracor.org/doc/api"
Repository = "https://github.com/dracor-org/dracor-mcp"

[tool.hatch.build.targets.wheel]
packages = ["."]
